% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_lim.r
\name{df2lim}
\alias{df2lim}
\title{Imports Declaration File as an lim object}
\usage{
df2lim(filename)
}
\arguments{
\item{filename}{A character string specifying the path to the declaration file to import.}
}
\value{
An lim object containing, among other components, the matrices A, B, G and H
that characterize the equality and inequality constraints defining the polytope 
associated to the LIM.
}
\description{
The function \code{df2lim()} turns a declaration file with linear inverse model specifications and creates the corresponding lim object.
}
\examples{
# Create a lim object from a Description file
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
}
