% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redpol.R
\name{lim.redpol}
\alias{lim.redpol}
\alias{red2full}
\alias{full2red}
\title{Projection of full polytope into the reduced polytope}
\usage{
lim.redpol(lim, test = TRUE)

red2full(sample, x0, Z)

full2red(sample, x0, Z)
}
\arguments{
\item{lim}{A list with four components \code{A}, \code{B}, \code{G} and \code{H} representing
the polytope to be reduced.}

\item{test}{A boolean if equal \code{TRUE} checks for additional equalities hidden in inequalities.}

\item{sample}{A matrix where each row corresponds to a point inside either the full or the reduced polytope.}

\item{x0}{A numeric vector of size \eqn{n} corresponding to the particular solution used during the reduction.}

\item{Z}{The matrix used during the reduction of the polytope (returned by \code{lim.redpol()}).}
}
\value{
A list with four components; namely:
\itemize{
  \item \code{G}
  \item \code{H}
  \item \code{x0}
  \item \code{Z}
}
}
\description{
The function \code{lim.redpol()} takes as input a polytope  and returns its projection into the non-empty reduced polytope.
Precisely, taking the polytope \eqn{\mathcal{P}= \{ x \in \mathbb{R}^n: Ax = B, Gx \geq H \}} as input, the function returns 
\itemize{
\item the matrix \code{Z}, basis of the right null space of A and \eqn{x_0} a particular solution of \eqn{P}, used for the reduction: a Chebyshev center, see \code{\link{pol.center}()};
\item the matrix \eqn{G'=GZ} and the vector \eqn{H'=H-Gx_0} describing the reduced polytope \eqn{\mathcal{P'}= \{ x \in \mathbb{R}^{n-k}: G'x \geq H'\}} with \eqn{k=\mathtt{rank}(A)}.
 }
 
 The function \code{full2red()} (resp. \code{red2full()}) turns a sample of points inside the full polytope \eqn{\mathcal{P}} (resp. inside the reduced polytope \eqn{\mathcal{P'}})
 into the sample of corresponding points inside the reduced polytope \eqn{\mathcal{P'}} (resp. the full polytope \eqn{\mathcal{P}}).
}
\examples{
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
BOWFred <- lim.redpol(BOWF)
str(BOWFred, max.length = 1)
}
