% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lim.R
\name{plot.lim}
\alias{plot.lim}
\title{Plot a metabolic network}
\usage{
\method{plot}{lim}(x, seed = NULL, ...)
}
\arguments{
\item{x}{An object of class `lim` containing a component `Flowmatrix`.}

\item{seed}{The seed for the PRNG of the R session}

\item{...}{Additionnal arguments to be passed to the function \code{plot.igraph()} from the package \code{\\{igraph\\}}.}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot a metabolic network
}
\examples{
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
plot(BOWF)
}
