% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_ranges.R
\name{pol.ranges}
\alias{pol.ranges}
\alias{lim.ranges}
\title{Determine ranges of a polytope}
\usage{
pol.ranges(A = NULL, B = NULL, G, H)

lim.ranges(lim)
}
\arguments{
\item{A}{A matrix corresponding to \code{A} in the description of the polytope \eqn{\mathcal{P}}.}

\item{B}{A numeric vector corresponding to \code{B} in the description of the polytope \eqn{\mathcal{P}}.}

\item{G}{A matrix corresponding to \code{G} in the description of the polytope \eqn{\mathcal{P}}.}

\item{H}{A numeric vector corresponding to \code{H} in the description of the polytope \eqn{\mathcal{P}}.}

\item{lim}{A list with four components \code{A}, \code{B}, \code{G} and \code{H} representing
the polytope.}
}
\value{
A \eqn{n \times 3} matrix, with \eqn{n} the dimension of the polytope. The three columns of the matrix correspond respectively
to the minimum, maximum and range along each dimension of the polytope.
}
\description{
The functions \code{pol.ranges()} and \code{lim.ranges()} compute the theoretical ranges of the polytope along each dimension of a given polytope \eqn{\mathcal{P}= \{ x \in \mathbb{R}^n: Ax = B, Gx \geq H \}}.
}
\examples{
# Create a lim object from a Description file
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
pol.ranges(A = BOWF$A, B = BOWF$B, G = BOWF$G, H = BOWF$H)
}
