% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplelim.R
\docType{package}
\name{samplelim-package}
\alias{samplelim}
\alias{samplelim-package}
\title{MCMC sampling algorithms for linear inverse models}
\description{
The package provides efficient implementations (C++ encoded) of 
Monte Carlo Markov Chains (MCMC) algorithms for uniformly sampling high-dimensional polytopes. 
It is particularly aimed at handling linear inverse models (LIM) in metabolic 
(trophic, biochemical or urban) networks. 
Some support functions are also included to facilitate its use by practitioners.
}
\references{
{
D. Van Oevelen, K. Van den Meersche, F. J. R. Meysman, K. Soetaert, 
J. J. Middelburg and A. F. Vézina, 
\emph{Quantifying Food Web Flows Using Linear Inverse Models},
 Ecosystems \strong{13}, 32-45 (2010).

B.T. Polyak and E.N. Gryazina, 
\emph{Billiard walk - a new sampling algorithm for control and optimization}, 
IFAC Proceedings Volumes, \strong{47(3)}, 6123-6128 (2014).

V. Girardin, T. Grente, N. Niquil and P. Regnault,
\emph{Comparing and updating R packages of MCMC Algorithms for 
Linear Inverse Modeling of Metabolic Networks}, to appear in
Communications in Statistics - Simulation and Computation (2025)
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pregnault/samplelim.git}
  \item Report bugs at \url{https://github.com/pregnault/samplelim/issues}
}

}
\author{
\strong{Maintainer}: Philippe Regnault \email{philipperegnault@hotmail.com} (\href{https://orcid.org/0000-0002-7086-9030}{ORCID})

Authors:
\itemize{
  \item Philippe Regnault \email{philipperegnault@hotmail.com} (\href{https://orcid.org/0000-0002-7086-9030}{ORCID})
  \item Jacques Bréhélin \email{breheli241@unicaen.fr}
  \item Matthieu Dien \email{matthieu.dien@unicaen.fr}
  \item Théo Grente \email{theo.grente@gmail.com}
}

Other contributors:
\itemize{
  \item Valérie Girardin \email{valerie.girardin@unicaen.fr} (\href{https://orcid.org/0000-0001-9934-3561}{ORCID}) [contributor]
  \item Nathalie Niquil \email{nathalie.niquil@unicaen.fr} [contributor]
  \item Quentin Noguès [data contributor]
}

}
\keyword{billard-walk}
\keyword{biochemical-network}
\keyword{lim}
\keyword{mcmc}
\keyword{metabolic-network}
\keyword{mirror-walk}
\keyword{polytope}
\keyword{sampling}
\keyword{trophic-network}
\keyword{urban-network}
