% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redpol.R
\name{lim.redpol}
\alias{lim.redpol}
\alias{red2full}
\alias{full2red}
\title{Projection of full polytope into the reduced polytope}
\usage{
lim.redpol(lim, test = TRUE)

red2full(sample, x0, Z)

full2red(sample, x0, Z)
}
\arguments{
\item{lim}{blabl}

\item{test}{pouet pouet}

\item{sample}{blabla}

\item{x0}{blabla}

\item{Z}{blabla}
}
\value{
A list with four components; namely:
\itemize{
  \item \code{G}
  \item \code{H}
  \item \code{x0}
  \item \code{Z}
}
}
\description{
Blabla description
}
\examples{
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
BOWFred <- lim.redpol(BOWF)
str(BOWFred, max.length = 1)
}
