% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol_ranges.R
\name{pol.ranges}
\alias{pol.ranges}
\alias{lim.ranges}
\title{Determine ranges of a polytope}
\usage{
pol.ranges(A = NULL, B = NULL, G, H)

lim.ranges(lim)
}
\arguments{
\item{A}{blabla}

\item{B}{blabla}

\item{G}{blabla}

\item{H}{blabla}

\item{lim}{blabla}
}
\value{
A length \eqn{n} numeric vector, with \eqn{n} the dimension of the polytope.
}
\description{
Blabla description
}
\examples{
# Create a lim object from a Description file
DF <- system.file("extdata", "DeclarationFileBOWF-short.txt", package = "samplelim")
BOWF <- df2lim(DF)
pol.ranges(A = BOWF$A, B = BOWF$B, G = BOWF$G, H = BOWF$H)
}
